requirements.txt
botogram2
pony
pyyaml